Major rework of monkeysphere-host to handle multiple host keys.
This rework removes any assumption that monkeysphere-host is just
managing a single host key, or that the keys are used specifically for
ssh.  The UI is exactly backwards compatible except that hostnames
('example.com') must be replaced by full service names
('ssh://example.com').  This incarnation passes the old tests with
those changes only.

There are a couple of things that still need to be done:

- need to see if a transition script is needed (some local file names
  have changed)

- need to fill in check_service_name function to verify that a
  specified service name fits the expected format.

- update diagnostics appropriately