fix ssh_proxycommand marginal ui
After the last big code cleanup, the bookkeeping of numbers of
processed/valid keys was removed.  This was done because most things
don't use that info, and it was confusing return codes of top-level
functions.

The one thing that did use that, though, was the ssh_proxycommand.  We
fix this by using a global variable to keep track of the number of
processed and valid keys.  The proxy command can now easily determine
when it should output it's marginal ui block.