Always consume stdin when log is called without a message argument
See the discussion about this in https://bugs.debian.org/635711

I'm hoping this will resolve the intermittent SIGPIPEs plaguing
monkeysphere's postinst.

I'm still unable to reproduce the problem.  If people who can
reproduce the problem could try this patch and report back if it fixes
things for them, that would be great.