Clean up some shell, coreutils, and mktemp loopholes
In some very unusual cases, where the system monkeysphere directory
has whitespace in the paths, or where some filenames start with a "-",
the shell, mktemp, or various components of coreutils could
misinterpret the commands given.

This cleans up some of those mistakes, many of which have been in the
monkeysphere codebase for ages.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>