Use 3072-bit RSA keys in tests
We were using 1024-bit RSA keys in tests to avoid blocking on systems
with limited entropy.  As we move into the getrandom() era, this is a
situation that we should expect the operating system to handle
correctly, and our tools should be tested with sensible limits and
sensible key sizes.