Avoid unnecessary user lockout
See the argument here: https://bugs.debian.org/897366

I discussed this on IRC with anarcat, and did some code archaeology,
and could not find any reason for the excessive paranoia.  If the
chown/chmod/mv sequence fails, then we should report an error and
remove the temporary incoming file, not the pre-existing one.